saftey updates

+++ nodeos/shutdown.sh
@@ -17,12 +17,14 @@ else
   exit
 fi
 
+cd ${LOG_DIR:?} || exit
+
 USER=$(id -un)
 PID=$(ps -u "$USER" | grep nodeos | cut -d" " -f2)
 # shutdown
 if [ -n "$PID" ]; then
-  echo "kill -15 $PID"
+  kill -15 $PID
 fi
 # clean out old logs
-find "$LOG_DIR" -mtime +14 | xargs /bin/rm -rf
+find "${LOG_DIR:?}" -mtime +14 | xargs /bin/rm -rf
 cp /dev/null "$LOG_DIR"/nodeos-eric-test.log
